+++ packages/eslint-config-angular/imports/index.js
@@ -1,9 +1,21 @@
 module.exports = {
   parser: '@typescript-eslint/parser',
-  plugins: ['@typescript-eslint', 'simple-import-sort', 'eslint-plugin-import'],
+  plugins: [
+    'import',
+    '@typescript-eslint',
+    'simple-import-sort',
+    'eslint-plugin-import',
+  ],
   settings: {
     'import/parsers': { '@typescript-eslint/parser': ['.ts'] },
-    'import/resolver': { 'eslint-import-resolver-typescript': true },
+    'import/resolver': {
+      'eslint-import-resolver-typescript': true,
+      typescript: {
+        // always try to resolve types under `<root>@types` directory
+        // even it doesn't contain any source code, like `@types/etc`
+        alwaysTryTypes: true,
+      },
+    },
   },
   rules: {
     'simple-import-sort/imports': 'error',
@@ -20,5 +32,8 @@ module.exports = {
     'import/no-duplicates': 'off',
     'no-duplicate-imports': 'off',
     '@typescript-eslint/no-duplicate-imports': 'error',
+
+    // overwrite rule from `@tinkoff/eslint-config`
+    '@typescript-eslint/consistent-type-imports': 'off',
   },
 };