+++ packages/eslint-config-angular/index.js
@@ -1,11 +1,3 @@
 module.exports = {
-  extends: [
-    './internal/base',
-    './internal/base-typescript',
-    './internal/html',
-    './internal/import',
-    './internal/member-ordering',
-    './internal/line-statements',
-    './internal/extraneous-class',
-  ],
+  extends: ['./recommended/index.js'],
 };