+++ packages/eslint-config-angular/member-ordering/index.js
@@ -2,6 +2,7 @@ module.exports = {
   overrides: [
     {
       files: ['*.ts'],
+      parser: '@typescript-eslint/parser',
       plugins: ['@typescript-eslint', 'sort-class-members'],
       rules: {
         /**
@@ -58,13 +59,13 @@ module.exports = {
                * [sort: public -> protected -> private]
                **/
               /*
-                'public-static-get',
-                'public-static-set',
-                'protected-static-get',
-                'protected-static-set',
-                'private-static-get',
-                'private-static-set',
-                */
+               * 'public-static-get',
+               * 'public-static-set',
+               * 'protected-static-get',
+               * 'protected-static-set',
+               * 'private-static-get',
+               * 'private-static-set',
+               */
 
               /**
                * @note(splincode):
@@ -76,19 +77,19 @@ module.exports = {
                * [sort: decorated -> non-decorated]
                **/
               /*
-                'public-decorated-get',
-                'public-instance-get',
-                'public-decorated-set',
-                'public-instance-set',
-                'protected-decorated-get',
-                'protected-instance-get',
-                'protected-decorated-set',
-                'protected-instance-set',
-                'private-decorated-get',
-                'private-instance-get',
-                'private-decorated-set',
-                'private-instance-set',
-                */
+               * 'public-decorated-get',
+               * 'public-instance-get',
+               * 'public-decorated-set',
+               * 'public-instance-set',
+               * 'protected-decorated-get',
+               * 'protected-instance-get',
+               * 'protected-decorated-set',
+               * 'protected-instance-set',
+               * 'private-decorated-get',
+               * 'private-instance-get',
+               * 'private-decorated-set',
+               * 'private-instance-set',
+               */
 
               /**
                * static methods