+++ packages/eslint-config-angular/recommended/base-typescript.js
@@ -4,6 +4,7 @@ module.exports = {
       files: ['*.ts'],
       parser: '@typescript-eslint/parser',
       plugins: ['@typescript-eslint'],
+      extends: ['plugin:@typescript-eslint/eslint-recommended'],
       parserOptions: {
         sourceType: 'module',
         errorOnUnknownASTType: true,
@@ -18,10 +19,23 @@ module.exports = {
          */
         'no-useless-constructor': 'off',
         '@typescript-eslint/no-useless-constructor': ['error'],
+        '@typescript-eslint/no-inferrable-types': [
+          'error',
+          { ignoreParameters: true },
+        ],
         '@typescript-eslint/prefer-readonly': ['error'],
         '@typescript-eslint/explicit-member-accessibility': [
           'error',
           { accessibility: 'no-public' },
+        ],
+        'no-unused-vars': 'off',
+        '@typescript-eslint/no-unused-vars': [
+          'error',
+          { argsIgnorePattern: '^_' },
+        ],
+        '@typescript-eslint/array-type': [
+          'error',
+          { default: 'array-simple', readonly: 'array-simple' },
         ],
       },
     },