+++ packages/eslint-config-angular/test/line-statements/line-statements-unhappy.test.js
@@ -6,7 +6,7 @@ describe('line statements / unhappy path', () => {
     cwd: path.join(__dirname, '..'),
     useEslintrc: false,
     baseConfig: {
-      extends: ['../internal/base', '../internal/line-statements'],
+      extends: ['../line-statements'],
     },
   });