+++ packages/eslint-config-angular/test/member-ordering/member-ordering-unhappy.test.js
@@ -6,7 +6,7 @@ describe('member-ordering / unhappy', () => {
     cwd: path.join(__dirname, '..'),
     useEslintrc: false,
     baseConfig: {
-      extends: ['../internal/base', '../internal/member-ordering'],
+      extends: ['../member-ordering'],
     },
   });