+++ packages/eslint-config/test/import/import-unhappy.test.js
@@ -1,19 +1,25 @@
 import ESlint from 'eslint';
 import path from 'path';
 
-describe('imports / happy path', () => {
+describe('import / unhappy path', () => {
   const cli = new ESlint.CLIEngine({
     cwd: path.join(__dirname, '..'),
     useEslintrc: false,
     baseConfig: {
-      extends: ['../imports'],
+      extends: ['../internal/import'],
+      parserOptions: {
+        sourceType: 'module',
+      },
+      env: {
+        es6: true,
+      },
     },
   });
 
-  it('happy', () => {
+  it('unhappy', () => {
     const codeframe = cli.getFormatter('codeframe');
     const report = cli.executeOnFiles([
-      path.join(__dirname, './__fixtures__/imports-happy.fixture.ts'),
+      path.join(__dirname, './__fixtures__/import-unhappy.fixture.js'),
     ]);
 
     expect(codeframe(report.results)).toMatchSnapshot();